pyproject.toml
[tool.poetry]
name = "klimaschutzmonitor"
version = "0.1.0"
description = "A website to maintain the progress of a municipality towards net zero."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
django-ordered-model = "^3.6"

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["cpmonitor"]
pythonVersion = "3.10"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pre-commit = "^3.0.4"
pyright = "^1.1.292"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
